backend/requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
supabase==2.4.6
pandas==2.2.2
